move btp name gen function to nginx provider
